requirements.txt
fastapi==0.111.0
fastapi-cli==0.0.4
uvicorn==0.30.1
uvloop==0.19.0
pydantic==2.7.4
pydantic_core==2.18.4
starlette==0.37.2
anyio==3.6.2
typing_extensions==4.12.2
celery==5.4.0
redis==4.5.5
redispy==3.0.0
kombu==5.3.7
billiard==4.2.0
amqp==5.2.0
vine==5.1.0
pymongo==4.9.1
motor==3.6.0
refile==8.0.3.post3
nori2==1.12.2
megfile==3.0.6.post1
oss2==2.19.0
crcmod==1.7
numpy==1.24.3
pandas==2.0.1
loguru==0.7.2
tqdm==4.66.4
python-multipart==0.0.9
python-statemachine==2.3.1
nacos-sdk-python==1.0.0
python-dotenv==1.0.1
brainpp==2.6.4
#MegBrain==8.9999.0+cu111
remisc==0.3.13
requests==2.30.0
requests-file==2.1.0
requests-futures==1.0.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
requests-unixsocket==0.2.0
urllib3==1.26.19
idna==2.8
certifi==2019.11.28
botocore==1.37.12
awscli==1.38.12
Pillow==9.5.0
opencv-python==4.7.0.72
lzf==0.1
ipython==8.12.2
ipython-genutils==0.2.0
pyquaternion==0.9.9
#aiobotocore~=2.4.0